requirements.txt
# ingestion
websocket-client>=1.5.1
kafka-python>=2.0.2

# core
pandas>=2.2
pyarrow>=15
python-dotenv>=1.0

# core
pandas>=2.2
pyarrow>=15
python-dotenv>=1.0

# analytics
duckdb>=0.10.0
google-cloud-bigquery>=3.20.0
google-cloud-storage>=2.16.0

# dbt
dbt-bigquery>=1.7,<1.9
dbt-duckdb>=1.7,<1.9
